20150131: Major revision, now includes error detection and a bunch of other cool stuff. Fixes a *LOT* of stuff!

+++ rlw.sh
@@ -19,93 +19,99 @@
 #
 #
 
-# Check that everything is here
-if [[ ! -e `which zenity` ]]; then
-	echo "Missing dependencies! Make sure zenity, wget, wine, and wine-staging are installed."
-	exit 1
-elif [[ ! -e `which wget` && `which wine` ]]; then
-	echo "Missing dependencies! Make sure zenity, wget, wine, and wine-staging are installed."
-	spawndialog error "Missing dependencies! Make sure zenity, \nwget, wine, and wine-staging are installed."
-	exit 1
-fi
+spawndialog () {
+	echo -e "$2"
+	zenity \
+		--window-icon="$RBXICON" \
+		--title='Roblox Linux Wrapper v'"$RLWVERSION"'-'"$RLWCHANNEL" \
+		--"$1" \
+		--no-wrap \
+		--text="$2"
+
+}
 
 # Define some variables and the spawndialog function
-export RLWVERSION=20150130-staging
-export RLWCHANNEL=RELEASE
+export RLWVERSION=20150131
+export RLWCHANNEL=staging
 export WINEARCH=win32
+
+# Check that everything is here
+[[ -e $(which zenity) && $(which wget) && $(which wine) ]] || { spawndialog error "Missing dependencies! Make sure zenity, wget, wine, and wine-staging are installed."; exit 1; }
+
+
 if [[ -e $HOME/.local/share/icons/hicolor/512x512/apps/roblox.png ]]; then
 	export RBXICON=$HOME/.local/share/icons/hicolor/512x512/apps/roblox.png
 fi
-echo 'Roblox Linux Wrapper v'$RLWVERSION'-'$RLWCHANNEL
-
-spawndialog () {
-	zenity \
-		--window-icon=$RBXICON \
-		--title='Roblox Linux Wrapper v'$RLWVERSION'-'$RLWCHANNEL \
-		--$1 \
-		--no-wrap \
-		--text="$2"
-}
+echo 'Roblox Linux Wrapper v'"$RLWVERSION"'-'"$RLWCHANNEL"
 
 # Uncomment these lines to use stock Wine (default)
-#export WINE=`which wine`
-#export WINESERVERBIN=`which wineserver`
-#export WINEPREFIX=$HOME/.local/share/wineprefixes/Roblox-wine
+#export WINE=$(which wine)
+#export WINEBOOTBIN=$(which wineboot)
+#export WINESERVERBIN=$(which wineserver)
+#export WINEPREFIX=$HOME/.local/share/wineprefixes/roblox-wine
+#export WINEPREFIX_OLD=$HOME/.local/share/wineprefixes/Roblox-wine
 
 # Uncomment these lines to use wine-staging (formerly wine-compholio)
 if [[ -f /opt/wine-compholio/bin/wine ]]; then
 	export WINE=/opt/wine-compholio/bin/wine
+	export WINEBOOTBIN=/opt/wine-compholio/bin/wineboot
 	export WINESERVERBIN=/opt/wine-compholio/bin/wineserver
-	export WINEPREFIX=$HOME/.local/share/wineprefixes/Roblox-wine-compholio
+	export WINEPREFIX=$HOME/.local/share/wineprefixes/roblox-wine-compholio
+	export WINEPREFIX_OLD=$HOME/.local/share/wineprefixes/Roblox-wine-compholio
 elif [[ -f /opt/wine-staging/bin/wine ]]; then
 	export WINE=/opt/wine-staging/bin/wine
+	export WINEBOOTBIN=/opt/wine-staging/bin/wineboot
 	export WINESERVERBIN=/opt/wine-staging/bin/wineserver
-	export WINEPREFIX=$HOME/.local/share/wineprefixes/Roblox-wine-staging
-else
-	echo "Missing dependencies! Make sure zenity, wget, wine, and wine-staging are installed."
-	spawndialog error "Missing dependencies! Make sure wine-staging is installed."
-	spawndialog question "Would you like to open the wine-staging installation instructions?"
-	if [[ $? == "0" ]]; then
-		xdg-open "https://github.com/wine-compholio/wine-staging/wiki/Installation"
-	fi
-	exit 1
+	export WINEPREFIX=$HOME/.local/share/wineprefixes/roblox-wine-staging
+	export WINEPREFIX_OLD=$HOME/.local/share/wineprefixes/Roblox-wine-staging
 fi
 
-download () {
-	wget $1 -O $2 2>&1 | \
-	sed -u 's/.* \([0-9]\+%\)\ \+\([0-9.]\+.\) \(.*\)/\1\n# Downloading at \2\/s, ETA \3/' | \
-	zenity \
-		--progress \
-		--window-icon=$RBXICON \
-		--title='Downloading' \
-		--auto-close \
-		--no-cancel \
-		--width=450 \
-		--height=120
+# Some internal functions to make wine more useful to the wrapper.
+# This allows the wrapper to know what went wrong and where, without excessive code.
+# Note: the "r" prefix indicates a function that extends system functionality.
+
+rwine () {
+	$WINE "$@"; [[ $? = "0" ]] || { spawndialog error "wine closed unsuccessfully.\nSee terminal for details. (exit code $?)"; exit $?; }
+}
+rwineboot () {
+	$WINEBOOTBIN; [[ $? = "0" ]] || { spawndialog error "wineboot closed unsuccessfully.\nSee terminal for details. (exit code $?)"; exit $?; }
+}
+rwineserver () {
+	$WINESERVERBIN "$@"; [[ $? = "0" ]] || { spawndialog error "wineserver closed unsuccessfully.\nSee terminal for details. (exit code $?)"; exit $?; }
+}
+rwget () {
+	wget "$1" -O "$2" 2>&1 | sed -u 's/.* \([0-9]\+%\)\ \+\([0-9.]\+.\) \(.*\)/\1\n# Downloading at \2\/s, ETA \3/' | zenity --progress --window-icon="$RBXICON" --title='Downloading' --auto-close --no-cancel --width=450 --height=120; [[ $? = "0" ]] || { spawndialog error "wget download failed. \nSee terminal for details. (exit code $?)"; exit $?; }
+}
+rwinetricks () {
+	$(which winetricks) "$@"
 }
 
 roblox-install () {
+	if [[ -e $WINEPREFIX_OLD ]] && [[ ! -e $WINEPREFIX ]]; then
+		mv "$WINEPREFIX_OLD" "$WINEPREFIX"
+	fi
 	if [[ ! -e $WINEPREFIX ]]; then
 		spawndialog question 'A working Roblox wineprefix was not found. Would you like to install one?'
 		if [[ $? == "0" ]]; then
 			# Make sure our directories really exist
-			mkdir -p $WINEPREFIX
-			download http://roblox.com/install/setup.ashx /tmp/RobloxPlayerLauncher.exe
-			download http://ftp.mozilla.org/pub/mozilla.org/firefox/releases/31.4.0esr/win32/en-US/Firefox%20Setup%2031.4.0esr.exe /tmp/Firefox-Setup-esr.exe
-			winetricks ddr=gdi
-			WINEDLLOVERRIDES="winebrowser.exe,winemenubuilder.exe=" $WINE /tmp/RobloxPlayerLauncher.exe
-			cd $WINEPREFIX
-			ROBLOXPROXY=`find . -iname 'RobloxProxy.dll' | sed "s/.\/drive_c/C:/" | tr '/' '\\'`
-			$WINE regsvr32 /i "$ROBLOXPROXY"
-			WINEDLLOVERRIDES="winebrowser.exe,winemenubuilder.exe=" $WINE /tmp/Firefox-Setup-esr.exe /SD | zenity \
-				--window-icon=$RBXICON \
+			[[ -e $HOME/.local/share/wineprefixes ]] || mkdir -p "$HOME/.local/share/wineprefixes"
+			rwineboot
+			[[ $? == 0 ]]  || { spawndialog error "Wine prefix not generated successfully.\nSee terminal for more details. (exit code $?)"; exit $?; }
+			rwget http://roblox.com/install/setup.ashx -O /tmp/RobloxPlayerLauncher.exe
+			rwget http://ftp.mozilla.org/pub/mozilla.org/firefox/releases/31.4.0esr/winwerwer32/en-US/Firefox%20Setup%2031.4.0esr.exe -O /tmp/Firefox-Setup-esr.exe
+			rwinetricks ddr=gdi
+			WINEDLLOVERRIDES="winebrowser.exe,winemenubuilder.exe=" rwine /tmp/RobloxPlayerLauncher.exe
+			cd "$WINEPREFIX"
+			ROBLOXPROXY="$(find . -iname 'RobloxProxy.dll' | sed "s/.\/drive_c/C:/" | tr '/' '\\')"
+			WINEDLLOVERRIDES="winebrowser.exe,winemenubuilder.exe=" rwine /tmp/Firefox-Setup-esr.exe /SD | zenity \
+				--window-icon="$RBXICON" \
 				--title='Installing Mozilla Firefox' \
 				--text='Installing Mozilla Firefox ESR ...' \
 				--progress \
 				--pulsate \
 				--no-cancel \
 				--auto-close
-			$WINESERVERBIN --wait
+			rwineserver --wait
 		else
 			exit 1
 		fi
@@ -115,7 +121,9 @@ roblox-install () {
 wrapper-install () {
 	if [[ ! -d $HOME/.rlw ]] || [[ ! -f $HOME/.local/share/applications/Roblox.desktop ]]; then
 		spawndialog question 'Roblox Linux Wrapper is not installed. This is necessary to launch games properly.\nWould you like to install it?'
-		if [[ $? == "0" ]]; then
+		if [[ $? == 0 ]]; then
+			[[ -f "$HOME/.local/share/icons/hicolor/512x512/apps/roblox.png" ]] || { mkdir -p "$HOME/.local/share/icons/hicolor/512x512/apps"; rwget http://img1.wikia.nocookie.net/__cb20130302012343/robloxhelp/images/f/fb/ROBLOX_Circle_Logo.png -O "$HOME/.local/share/icons/hicolor/512x512/apps/roblox.png"; }
+			export RBXICON=$HOME/.local/share/icons/hicolor/512x512/apps/roblox.png
 			cat <<-EOF > $HOME/.local/share/applications/Roblox.desktop
 			[Desktop Entry]
 			Comment=Play Roblox
@@ -135,21 +143,16 @@ wrapper-install () {
 			Name='Roblox for All'
 			Exec=xdg-open 'http://www.roblox.com/Groups/group.aspx?gid=292611'
 			EOF
-			mkdir -p $HOME/.rlw
-			download https://raw.githubusercontent.com/alfonsojon/roblox-linux-wrapper/master/rlw.sh $HOME/.rlw/rlw.sh
-			download https://raw.githubusercontent.com/alfonsojon/roblox-linux-wrapper/master/rlw-stub.sh $HOME/.rlw/rlw-stub.sh
-			download http://img1.wikia.nocookie.net/__cb20130302012343/robloxhelp/images/f/fb/ROBLOX_Circle_Logo.png $HOME/.local/share/icons/roblox.png
-			chmod +x $HOME/.rlw/rlw.sh
-			chmod +x $HOME/.rlw/rlw-stub.sh
-			chmod +x $HOME/.local/share/applications/Roblox.desktop
-			xdg-desktop-menu install --novendor $HOME/.local/share/applications/Roblox.desktop
+			mkdir -p "$HOME/.rlw"
+			rwget https://raw.githubusercontent.com/alfonsojon/roblox-linux-wrapper/master/rlw.sh -O "$HOME/.rlw/rlw.sh"
+			rwget https://raw.githubusercontent.com/alfonsojon/roblox-linux-wrapper/master/rlw-stub.sh -O "$HOME/.rlw/rlw-stub.sh"
+			rwget http://img1.wikia.nocookie.net/__cb20130302012343/robloxhelp/images/f/fb/ROBLOX_Circle_Logo.png -O "$HOME/.local/share/icons/roblox.png"
+			chmod +x "$HOME/.rlw/rlw.sh"
+			chmod +x "$HOME/.rlw/rlw-stub.sh"
+			chmod +x "$HOME/.local/share/applications/Roblox.desktop"
+			xdg-desktop-menu install --novendor "$HOME/.local/share/applications/Roblox.desktop"
 			xdg-desktop-menu forceupdate
-			if [[ -f $HOME/.rlw/rlw-stub.sh && -f $HOME/.rlw/rlw.sh && -f $HOME/.local/share/icons/roblox.png && -f $HOME/.local/share/applications/Roblox.desktop ]]; then
-				echo "wrapper installed properly, continuing"
-			else
-				spawndialog error 'Roblox Linux Wrapper did not install successfully.'
-				exit 1
-			fi
+			[[ -f $HOME/.rlw/rlw-stub.sh && -f $HOME/.rlw/rlw.sh && -f $HOME/.local/share/icons/roblox.png && -f $HOME/.local/share/applications/Roblox.desktop ]] || { spawndialog error 'Roblox Linux Wrapper did not install successfully.'; exit 1; }
 		else
 			exit 1
 		fi
@@ -157,41 +160,36 @@ wrapper-install () {
 }
 
 playerwrapper () {
-	ROBLOXPROXY=`find . -iname 'RobloxProxy.dll' | sed "s/.\/drive_c/C:/" | tr '/' '\\'`
-	$WINE regsvr32 /i "$ROBLOXPROXY"
+	ROBLOXPROXY=$(find . -iname 'RobloxProxy.dll' | sed "s/.\/drive_c/C:/" | tr '/' '\\')
+	rwine regsvr32 /i "$ROBLOXPROXY"
 	if [[ $1 = legacy ]]; then
-		export GAMEURL=`\
-		zenity \
-			--title='Roblox Linux Wrapper v'$RLWVERSION'-'$RLWCHANNEL \
-			--window-icon=$RBXICON \
-			--entry \
-			--text='Paste the URL for the game here.' \
-			--ok-label='Play' \
-			--width=450 \
-			--height=122`
-			GAMEID=`echo $GAMEURL | cut -d "=" -f 2`
+		export GAMEURL=$(\
+			zenity \
+				--title='Roblox Linux Wrapper v'$RLWVERSION'-'$RLWCHANNEL \
+				--window-icon="$RBXICON" \
+				--entry \
+				--text='Paste the URL for the game here.' \
+				--ok-label='Play' \
+				--width=450 \
+				--height=122)
+			GAMEID=$(echo "$GAMEURL" | cut -d "=" -f 2)
 		if [[ -n "$GAMEID" ]]; then
-			$WINE "`find $WINEPREFIX -name RobloxPlayerBeta.exe`" --id $GAMEID
-			$WINESERVERBIN --wait
+			rwine "$(find "$WINEPREFIX" -name RobloxPlayerBeta.exe)" --id "$GAMEID"
+			rwineserver --wait
 		else
 			return
 		fi
 	else
-		$WINE 'C:\Program Files\Mozilla Firefox\firefox.exe' http://www.roblox.com/Games.aspx
+		rwine 'C:\Program Files\Mozilla Firefox\firefox.exe' http://www.roblox.com/Games.aspx
 	fi
 }
 
 main () {
-	if [[ ! -f $HOME/.local/share/icons/hicolor/512x512/apps/roblox.png ]]; then
-		mkdir -p $HOME/.local/share/icons/hicolor/512x512/apps
-		download http://img1.wikia.nocookie.net/__cb20130302012343/robloxhelp/images/f/fb/ROBLOX_Circle_Logo.png $HOME/.local/share/icons/hicolor/512x512/apps/roblox.png
-		export RBXICON=$HOME/.local/share/icons/hicolor/512x512/apps/roblox.png
-	fi
-	rm -rf $HOME/Desktop/ROBLOX*desktop $HOME/Desktop/ROBLOX*.lnk
-	rm -rf $HOME/.local/share/applications/wine/Programs/Roblox
-	sel=`zenity \
+	rm -rf "$HOME/Desktop/ROBLOX*desktop $HOME/Desktop/ROBLOX*.lnk"
+	rm -rf "$HOME/.local/share/applications/wine/Programs/Roblox"
+	sel=$(zenity \
 		--title='Roblox Linux Wrapper v'$RLWVERSION'-'$RLWCHANNEL' by alfonsojon' \
-		--window-icon=$RBXICON \
+		--window-icon="$RBXICON" \
 		--width=480 \
 		--height=240 \
 		--cancel-label='Quit' \
@@ -204,35 +202,29 @@ main () {
 		FALSE 'Play Roblox (Legacy Mode)' \
 		FALSE 'Roblox Studio' \
 		FALSE 'Reset Roblox to defaults' \
-		FALSE 'Uninstall Roblox' `
+		FALSE 'Uninstall Roblox')
 	case $sel in
 	'Play Roblox')
 		playerwrapper; main;;
 	'Play Roblox (Legacy Mode)')
 		playerwrapper legacy; main;;
 	'Roblox Studio')
-		if [[ `find $WINEPREFIX/drive_c/users/$USER/Local\ Settings/Application\ Data/RobloxVersions/version-* -name RobloxStudioLauncherBeta.exe` == '' ]]; then
-			$WINE $WINEPREFIX/drive_c/users/$USER/Local\ Settings/Application\ Data/RobloxVersions/RobloxStudioLauncherBeta.exe -ide
-			$WINESERVERBIN -k
-		fi
-		WINEDLLOVERRIDES="msvcp110.dll,msvcr110.dll=n,b" $WINE $WINEPREFIX/drive_c/users/$USER/Local\ Settings/Application\ Data/RobloxVersions/version-*/RobloxStudioBeta.exe
-		$WINESERVERBIN --wait
+		WINEDLLOVERRIDES="msvcp110.dll,msvcr110.dll=n,b" rwine "$WINEPREFIX/drive_c/users/$USER/Local Settings/Application Data/RobloxVersions/RobloxStudioLauncherBeta.exe" -ide
+		rwineserver --wait
 		main;;
 	'Reset Roblox to defaults')
-		rm -rf $WINEPREFIX;
+		rm -rf "$WINEPREFIX";
 		roblox-install; main;;
 	'Uninstall Roblox')
 		spawndialog question 'Are you sure you would like to uninstall?'
 		if [[ $? == "0" ]]; then
-			xdg-desktop-menu uninstall $HOME/.local/share/applications/Roblox.desktop
-			rm -rf $HOME/.rlw
-			if [[ -e $HOME/.local/share/icons/roblox.png ]]; then
-				rm -rf $HOME/.local/share/icons/roblox.png
-			fi
-			rm -rf $HOME/.local/share/icons/hicolor/512x512/apps/roblox.png
+			xdg-desktop-menu uninstall "$HOME/.local/share/applications/Roblox.desktop"
+			rm -rf "$HOME/.rlw"
+			[[ ! -e "$HOME/.local/share/icons/roblox.png" ]] || rm -rf "$HOME/.local/share/icons/roblox.png"
+			rm -rf "$HOME/.local/share/icons/hicolor/512x512/apps/roblox.png"
 			xdg-desktop-menu forceupdate
 			$WINESERVERBIN --kill
-			rm -rf $WINEPREFIX
+			rm -rf "$WINEPREFIX"
 			if [[ -d $HOME/.rlw ]] || [[ -e $HOME/.local/share/icons/hicolor/512x512/apps/roblox.png ]] || [[ -d $WINEPREFIX ]]; then
 				spawndialog error 'Roblox is still installed. Please try uninstalling again.'
 			else
